+++ docs/astro.config.ts
@@ -1,21 +1,9 @@
 import { defineConfig, envField } from 'astro/config';
 import starlight from '@astrojs/starlight';
-import type { StarlightConfig } from '@astrojs/starlight/types';
 import vercel from '@astrojs/vercel';
 import starWarp from '@inox-tools/star-warp';
 import starlightLinksValidator from 'starlight-links-validator';
 import starlightCoolerCredit from 'starlight-cooler-credit';
-
-const badge = {
-	new: {
-		text: 'NEW',
-		variant: 'success',
-	},
-	updated: {
-		text: 'UPDATED',
-		variant: 'default',
-	},
-} satisfies Record<string, NonNullable<NonNullable<StarlightConfig['sidebar']>[number]['badge']>>;
 
 process.env.ASTRO_PROJECT_ROOT = new URL('../', import.meta.url).toString();
 
@@ -106,9 +94,13 @@ export default defineConfig({
 						{
 							label: 'Content Utilities',
 							collapsed: false,
-							autogenerate: {
-								directory: 'content-utils',
-							},
+							items: [
+								{
+									autogenerate: {
+										directory: 'content-utils',
+									},
+								},
+							],
 						},
 					],
 				},
@@ -127,14 +119,18 @@ export default defineConfig({
 						{
 							label: 'Modular Station',
 							collapsed: false,
-							autogenerate: {
-								directory: 'modular-station',
-							},
+							items: [
+								{
+									autogenerate: {
+										directory: 'modular-station',
+									},
+								},
+							],
 						},
 						{
 							label: 'Inline Module',
 							collapsed: true,
-							autogenerate: { directory: 'inline-mod' },
+							items: [{ autogenerate: { directory: 'inline-mod' } }],
 						},
 					],
 				},